Added initial bank functions

+++ src/azure-function/Models/Account.cs
@@ -4,5 +4,5 @@ public class Account
 {
     public string AccountNumber { get; set; }
     public string AccountName { get; set; }
-    public decimal AccountBalance { get; set; }
+    public double AccountBalance { get; set; }
 }
